+++ docs/apps/schedy/CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 * Schedy can now re-evaluate schedules automatically when the state of
   entities changes. See the new ``watched_entities`` settings.
+* Range specifications for constraints can now be inverted by prepending
+  them with ``!``.
 
 ### Changed
 * Various small improvements of the examples for using expressions.